+++ app/src/main/java/com/github/sasd97/android_router/FirstFragment.java
@@ -1,0 +1,21 @@
+package com.github.sasd97.android_router;
+
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+/**
+ * Created by alexander on 06/09/2017.
+ */
+
+public class FirstFragment extends Fragment {
+
+    @Nullable
+    @Override
+    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.fragment_first, container, false);
+    }
+}
