+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="sasd97.java_blog.xyz.android_router.MainActivity">
+    tools:context="com.github.sasd97.android_router.MainActivity">
 
     <Button
         android:id="@+id/next"
