+++ lib-router/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
 
-    package="sasd97.java_blog.xyz.lib_router">
+    package="com.github.sasd97.lib_router">
 
     <application android:allowBackup="true" android:label="@string/app_name"
         android:supportsRtl="true">
