+++ lib-router/src/main/java/com/github/sasd97/lib_router/commands/Command.java
@@ -1,0 +1,8 @@
+package com.github.sasd97.lib_router.commands;
+
+/**
+ * Created by alexander on 06/09/2017.
+ */
+
+public interface Command {
+}
