+++ lib-router/src/main/java/com/github/sasd97/lib_router/exceptions/CommandNotSupportedException.java
@@ -1,0 +1,12 @@
+package com.github.sasd97.lib_router.exceptions;
+
+/**
+ * Created by alexander on 06/09/2017.
+ */
+
+public final class CommandNotSupportedException extends RuntimeException {
+
+    public CommandNotSupportedException(String message) {
+        super(message);
+    }
+}
