--- lib-router/src/main/java/sasd97/java_blog/xyz/lib_router/commands/Command.java
@@ -1,8 +1,0 @@
-package sasd97.java_blog.xyz.lib_router.commands;
-
-/**
- * Created by alexander on 06/09/2017.
- */
-
-public interface Command {
-}
