--- lib-router/src/main/java/sasd97/java_blog/xyz/lib_router/commands/CommandDecorator.java
@@ -1,19 +1,0 @@
-package sasd97.java_blog.xyz.lib_router.commands;
-
-import android.support.annotation.NonNull;
-
-/**
- * Created by alexander on 06/09/2017.
- */
-
-public abstract class CommandDecorator<T extends Command> implements Command {
-
-    protected T command;
-
-    public CommandDecorator() {
-    }
-
-    public CommandDecorator(@NonNull T command) {
-        this.command = command;
-    }
-}
