--- lib-router/src/main/java/sasd97/java_blog/xyz/lib_router/exceptions/CommandNotSupportedException.java
@@ -1,12 +1,0 @@
-package sasd97.java_blog.xyz.lib_router.exceptions;
-
-/**
- * Created by alexander on 06/09/2017.
- */
-
-public final class CommandNotSupportedException extends RuntimeException {
-
-    public CommandNotSupportedException(String message) {
-        super(message);
-    }
-}
