--- lib-router/src/main/java/sasd97/java_blog/xyz/lib_router/satellites/FragmentSatellite.java
@@ -1,34 +1,0 @@
-package sasd97.java_blog.xyz.lib_router.satellites;
-
-import android.support.annotation.NonNull;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentTransaction;
-
-import sasd97.java_blog.xyz.lib_router.commands.Command;
-import sasd97.java_blog.xyz.lib_router.commands.fragments.FragmentCommand;
-
-/**
- * Created by alexander on 06/09/2017.
- */
-
-public final class FragmentSatellite implements Satellite {
-
-    private int containerId;
-    private FragmentManager fragmentManager;
-
-    public FragmentSatellite(int containerId,
-                             @NonNull FragmentManager fragmentManager) {
-        this.containerId = containerId;
-        this.fragmentManager = fragmentManager;
-    }
-
-    @Override
-    public void execute(@NonNull Command command) {
-        if (!(command instanceof FragmentCommand)) return;
-
-        FragmentTransaction transaction = fragmentManager.beginTransaction();
-        transaction = ((FragmentCommand) command).apply(containerId, transaction);
-        transaction.commit();
-    }
-}
